+++ src/ResourceManager/Resources/Commands.Resources/RoleAssignments/NewAzureRoleAssignmentCommand.cs
@@ -59,8 +59,8 @@ namespace Microsoft.Azure.Commands.Resources
         [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true, ParameterSetName = ParameterSet.ScopeWithSPN,
             HelpMessage = "The app SPN.")]
         [ValidateNotNullOrEmpty]
-        [Alias("SPN")]
-        public string ServicePrincipalName { get; set; }
+        [Alias("SPN","ServicePrincipalName")]
+        public string ApplicationId { get; set; }
 
         [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true, ParameterSetName = ParameterSet.ResourceGroupWithObjectId,
             HelpMessage = "Resource group to assign the role to.")]
@@ -160,7 +160,7 @@ namespace Microsoft.Azure.Commands.Resources
                 ADObjectFilter = new ADObjectFilterOptions
                 {
                     UPN = SignInName,
-                    SPN = ServicePrincipalName,
+                    SPN = ApplicationId,
                     Id = ObjectId == Guid.Empty ? null : ObjectId.ToString(),
                 },
                 ResourceIdentifier = new ResourceIdentifier()