+++ src/ResourceManager/Resources/ChangeLog.md
@@ -29,6 +29,8 @@
 * Allow users to create RoleAssignment using delegation flag
     - Users can now use New-AzureRMRoleAssignment with an option of adding the delegation flag
 * Fix RoleAssignment get to respect the scope parameter
+* Add an alias for ServicePrincipalName in the New-AzureRmRoleAssignment Commandlet
+    - Users can now use the ApplicationId instead of the ServicePrincipalName when using the New-AzureRmRoleAssignment commandlet
 
 ## Version 5.0.0
 * Add support for online help